requirements.txt
numpy==1.24.3
pandas==1.5.3
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.13
langchain==0.1.12
openai==1.13.3
tiktoken==0.6.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1
youtube-transcript-api==0.6.2
deep-translator==1.11.4
langdetect==1.0.9
newspaper3k==0.2.8
readability-lxml==0.8.1
cloudscraper==1.2.71
tqdm==4.66.2